requirements.txt
SpeechRecognition
pydub
pyannote.audio
python-dotenv
torch
torchaudio
